moving all soundfont logic to audio library
